Moved WidgetWrapper to components
