+++ client/scripts/index.jsx
@@ -5,7 +5,7 @@ import thunkMiddleware from 'redux-thunk';
 import { Provider } from 'react-redux';
 import Application from './containers/App';
 import { BOAArtrosSkola } from './widgets';
-import WidgetWrapper from './widgets/WidgetWrapper'
+import WidgetWrapper from './components/WidgetWrapper'
 import loginApp from './reducers/reducers';
 
 // Webpack dependencies
